manual changes to handle database refactoring for shared usage
